drm/amd/dal: Only create up to the max number of targets for asic
This change limits the number of crtc's reported to drm to the
maximum number of displays that can be driven.

This fixes pageflip/validation issues when trying to drive
more displays than capable in the hardware. Usermode continously
failed validation since it keeps trying to lower the timing to fit
within the max bandwidth, but validation will always fail in this case
since it fails due to not having enough controllers, not due to
bandwidth restraints

Acked-by: Harry Wentland <harry.wentland@amd.com>

+++ drivers/gpu/drm/amd/dal/dc/inc/core_types.h
@@ -239,8 +239,8 @@ struct resource_pool {
 	struct timing_generator *timing_generators[MAX_PIPES];
 	struct stream_encoder *stream_enc[MAX_PIPES * 2];
 
-	uint8_t pipe_count;
-	uint8_t stream_enc_count;
+	unsigned int pipe_count;
+	unsigned int stream_enc_count;
 
 	union supported_stream_engines stream_engines;
 
@@ -250,10 +250,10 @@ struct resource_pool {
 	struct clock_source *dp_clock_source;
 
 	struct clock_source *clock_sources[MAX_CLOCK_SOURCES];
-	uint8_t clk_src_count;
+	unsigned int clk_src_count;
 
 	struct audio *audios[MAX_PIPES];
-	uint8_t audio_count;
+	unsigned int audio_count;
 
 	struct display_clock *display_clock;
 	struct adapter_service *adapter_srv;
